Overhaul data export support added in 8e7add27401e6ff410baa91
# Performance improvements

The previous mechanism was extremely heavy handed in terms of how it
exported data off the site: it would previously download the raw
activity JSON just for the activity name and start date, then throw
the rest of the activity data away. This wasted a lot of data (on the
order of MB for larger activities) per export operation and added a lot
more network overhead than was really needed.

Take the data from the activity metadata, i.e., the `.get_activities()`
derived values, and apply it to the exported activities instead.

# Timeout modifications

Adjust timeouts to better handle raw activity query API calls: now
timeouts are split between simple API calls and the raw activity query
API calls.

# Limit exporting positional data to supporting activities

Only select activities are available for export via GPX and KML file
format as not all activity types, e.g., jumping rope, elliptical,
gym cardio, etc, don't have positional data associated with them.
There's no sense hitting the server if we know the activity type does
not have positional data associated with it.

# Expose query limits at the data extractor level

Prior to this change all exports were stopped at the 200 entries mark.
This was happening because the limit could not be adjusted from the
export function as it consumed the data provided by the
`.get_activities()` function. This makes it easier to tune the
behavior without having to resort to specifying the limit in each
high-level API call provided by the data module.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>